+++ src/js/components/scrollTo.js
@@ -11,6 +11,14 @@ function scroll(element, to, duration) {
   }, 10);
 }
 function scrollTo (to,duration) {
+  var wh = k.win.H(),
+      bh = k('body').outerHeight(),
+      total = to + wh,
+      goal = bh - wh - 1;
+  if (total > bh) { to = goal; };
+  
+  to = Math.round(to);
+
   scroll(document.body, to, duration);
   scroll(document.documentElement, to, duration);
 }