+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### v2.4.0
 + Add mixin `slider-carousel`, `slider-gallery`
++ Add js function `equalizer`, `sticky`
 
 ### v2.3.0
 + Add mixin `tabs`, `tabs-active`, `tabs-panel-active`.